+++ CovidSafe/CovidSafe.API/v1/Controllers/MessageControllers/SizeController.cs
@@ -5,17 +5,18 @@ using System.Threading.Tasks;
 
 using CovidSafe.DAL.Services;
 using CovidSafe.Entities.Protos;
+using CovidSafe.Entities.Protos.Deprecated.v1;
 using CovidSafe.Entities.Validation;
 using Microsoft.AspNetCore.Http;
 using Microsoft.AspNetCore.Mvc;
 
-namespace CovidSafe.API.Controllers.MessageControllers
+namespace CovidSafe.API.v1.Controllers.MessageControllers
 {
     /// <summary>
     /// Handles <see cref="MatchMessage"/> size requests
     /// </summary>
     [ApiController]
-    [ApiVersion("1")]
+    [ApiVersion("1", Deprecated = true)]
     [Route("api/v{version:apiVersion}/Messages/[controller]")]
     public class SizeController : ControllerBase
     {
