+++ CovidSafe/CovidSafe.Entities/Protos/Deprecated/v1/MessageSizeRequest.cs
@@ -1,8 +1,24 @@
 ﻿using CovidSafe.Entities.Validation;
 using CovidSafe.Entities.Validation.Resources;
 
-namespace CovidSafe.Entities.Protos
+#pragma warning disable CS1591, CS0612, CS3021, IDE1006
+namespace CovidSafe.Entities.Protos.Deprecated.v1
 {
+    [global::ProtoBuf.ProtoContract()]
+    public partial class MessageSizeRequest : global::ProtoBuf.IExtensible
+    {
+        private global::ProtoBuf.IExtension __pbn__extensionData;
+        global::ProtoBuf.IExtension global::ProtoBuf.IExtensible.GetExtensionObject(bool createIfMissing)
+            => global::ProtoBuf.Extensible.GetExtensionObject(ref __pbn__extensionData, createIfMissing);
+
+        [global::ProtoBuf.ProtoMember(1, Name = @"region")]
+        public Region Region { get; set; }
+
+        [global::ProtoBuf.ProtoMember(2, Name = @"last_query_time")]
+        public long LastQueryTime { get; set; }
+
+    }
+
     /// <summary>
     /// <see cref="MessageSizeRequest"/> partial from generated Protobuf class
     /// </summary>
@@ -35,3 +51,4 @@ namespace CovidSafe.Entities.Protos
         }
     }
 }
+#pragma warning restore CS1591, CS0612, CS3021, IDE1006
