+++ CovidSafe/CovidSafe.Entities/Protos/Interactions.proto
@@ -3,17 +3,6 @@ syntax = "proto3";
 package Corona;
 option csharp_namespace = "CovidSafe.Entities.Protos";
 
-// Phone -> Server; Request to get size of messages queued for a region since last query time
-message MessageSizeRequest {
-  Region region = 1;
-  int64 last_query_time = 2;
-}
-
-// Server -> Phone; Provide size of query in bytes of the response
-message MessageSizeResponse {
-  int64 size_of_query_response = 1;
-}
-
 // Phone -> Server;
 // Get list of messages queued for a region since last query time
 message MessageListRequest {
@@ -25,7 +14,7 @@ message MessageListRequest {
 // as_of tells a client the current server time to enable clock synchronization
 message MessageListResponse {
   repeated MessageInfo message_info = 1;
-  int64 as_of = 2;
+  int64 max_response_timestamp = 2;
 }
 
 // Phone -> Server GetMessages(new_message_ids)
@@ -64,8 +53,6 @@ message SelfReportRequest {
   // Coarse region that this request applies to specified
   // to maximum allowed privacy preserving precision
   Region region = 2;
-  // Current timestamp on the client, in UTC ms since UNIX epoch
-  int64 client_timestamp = 3;
 }
 
 // Server -> Phone
@@ -104,10 +91,6 @@ message BlueToothSeed{
   string seed = 1;
   int64 sequence_start_time = 2;
   int64 sequence_end_time = 3;
-  // This is calculated on the server-side, based on the difference between the 
-  // submitted client_timestamp in SelfReportRequest and the current server time
-  // In milliseconds
-  int64 estimated_skew = 4;
 }
 
 message Area {
